packages/auth/pyproject.toml
[project]
name = "snip-auth"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "python-jose[cryptography]",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-cov", "ruff", "pyright>=1.1.400"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "UP006",
    "UP007",
    "UP035",
    "UP045",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
exclude = ["tests"]
